Fixed homework errors, added chat room stuff, darkmode route and dropdown

+++ routes/chat.js
@@ -61,6 +61,11 @@ router.get('/:id', middleware.isLoggedIn, middleware.checkIfMember, (req, res) =
       await room.save();
     }
 
+    if (req.user.newRoomCount.includes(room._id)) {
+      req.user.newRoomCount.splice(req.user.newRoomCount.indexOf(room._id), 1);
+      await req.user.save();
+    }
+
     const comments = await Comment.find({ room: room._id }).sort({ _id: -1 }).limit(30)
     .populate({path: 'author', select: ['username', 'imageUrl']});
 
@@ -98,18 +103,20 @@ router.get('/:id/edit', middleware.isLoggedIn, middleware.checkRoomOwnership, (r
 
 // create new rooms
 router.post('/', middleware.isLoggedIn, (req, res) => {
-  const room = {
-    name: filter.clean(req.body.name),
-    'creator.id': req.user._id,
-    'creator.username': req.user.username,
-    members: [req.user._id]
-  };
+  (async() => {
 
-  Room.create(room, (err, room) => {
-    if (err) {
-      console.log(err);
-      req.flash('error', 'group could not be created');
-      res.redirect('/chat/new');
+    const roomObject = {
+      name: filter.clean(req.body.name),
+      'creator.id': req.user._id,
+      'creator.username': req.user.username,
+      members: [req.user._id]
+    };
+
+    const room = await Room.create(roomObject);
+    if (!room) {
+      req.flash('error', 'Group could not be created');
+      return res.redirect('/chat/new');
+
     } else {
       if(req.body.type == 'true') {
         for (const user in req.body.check) {
@@ -123,11 +130,25 @@ router.post('/', middleware.isLoggedIn, (req, res) => {
       if(req.body.description) {
         room.description = filter.clean(req.body.description);
       }
-      room.save();
-      console.log('Database Room created: '.cyan);
-      console.log(room);
-      res.redirect('/chat/' + room._id);
+
+      await room.save();
+
+      const members = await User.find({_id: {$in: room.members}});
+      if (!members) {
+        req.flash('error', 'Members could not be found');
+        return res.redirect('/chat/new');
+      }
+
+      for (let member of members) {
+        member.newRoomCount.push(room._id);
+        member.save();
+      }
+
+      return res.redirect('/chat/' + room._id);
     }
+  })().catch(err => {
+    req.flash('error', 'Group could not be created');
+    res.redirect('/chat/new');
   });
 });
 
@@ -141,6 +162,11 @@ router.post('/:id/leave', middleware.isLoggedIn, middleware.checkForLeave, (req,
 
     if(room.creator.id.equals(req.user._id)) {
       req.flash('error', 'You cannot leave a room you created');
+      return res.redirect('back');
+    }
+
+    if (!room.mutable) {
+      req.flash('error', 'You cannot leave this room');
       return res.redirect('back');
     }
 
@@ -180,6 +206,10 @@ router.post('/:id/request-access', middleware.isLoggedIn, (req, res) => {
 
     if(foundRoom.type == 'public') {
       req.flash('error', 'Room is public');
+      return res.redirect('back');
+
+    } else if (!foundRoom.mutable) {
+      req.flash('error', 'Room does not accept access requests');
       return res.redirect('back');
     }
 
@@ -290,13 +320,31 @@ router.put('/:id', middleware.isLoggedIn, middleware.checkRoomOwnership, (req, r
 // delete room
 router.delete('/:id/delete', middleware.isLoggedIn, middleware.checkRoomOwnership, (req, res) => {
   (async () => {
-    const deletedRoom = await Room.findByIdAndDelete(req.params.id);
+
+    const room = await Room.findById(req.params.id);
+    if (!room) {
+      req.flash('error', 'A problem occured');
+      return res.redirect('back');
+    }
+
+    if (!room.mutable) {
+      req.flash('error', 'You cannot delete this room');
+      return res.redirect('back');
+    }
+
+    const deletedRoom = await Room.findByIdAndDelete(req.params.id).populate("members");
     if(!deletedRoom) {req.flash('error', 'A problem occured'); return res.redirect('back');}
 
     await Promise.all([
       Comment.deleteMany({room: deletedRoom._id}),
       AccessReq.deleteMany({room: deletedRoom._id})
     ]);
+
+    for (let member of deletedRoom.members) {
+      if (member.newRoomCount.includes(deletedRoom._id)) {
+        member.newRoomCount.splice(member.newRoomCount.indexOf(deletedRoom._id), 1);
+      }
+    }
 
     req.flash('success', 'Deleted room');
     res.redirect('/chat');
